fix: use decimal for tpch benchmark

+++ benchmark/clickbench/tpch/create_local.sql
@@ -4,7 +4,7 @@ CREATE TABLE IF NOT EXISTS customer (
     c_address STRING not null,
     c_nationkey INTEGER not null,
     c_phone STRING not null,
-    c_acctbal DOUBLE not null,
+    c_acctbal DECIMAL(15, 2) not null,
     c_mktsegment STRING not null,
     c_comment STRING not null
 ) storage_format = 'native' compression = 'lz4';
@@ -13,10 +13,10 @@ CREATE TABLE IF NOT EXISTS lineitem (
     l_partkey BIGINT not null,
     l_suppkey BIGINT not null,
     l_linenumber BIGINT not null,
-    l_quantity DOUBLE not null,
-    l_extendedprice DOUBLE not null,
-    l_discount DOUBLE not null,
-    l_tax DOUBLE not null,
+    l_quantity DECIMAL(15, 2) not null,
+    l_extendedprice DECIMAL(15, 2) not null,
+    l_discount DECIMAL(15, 2) not null,
+    l_tax DECIMAL(15, 2) not null,
     l_returnflag STRING not null,
     l_linestatus STRING not null,
     l_shipdate DATE not null,
@@ -36,7 +36,7 @@ CREATE TABLE IF NOT EXISTS orders (
     o_orderkey BIGINT not null,
     o_custkey BIGINT not null,
     o_orderstatus STRING not null,
-    o_totalprice DOUBLE not null,
+    o_totalprice DECIMAL(15, 2) not null,
     o_orderdate DATE not null,
     o_orderpriority STRING not null,
     o_clerk STRING not null,
@@ -47,7 +47,7 @@ CREATE TABLE IF NOT EXISTS partsupp (
     ps_partkey BIGINT not null,
     ps_suppkey BIGINT not null,
     ps_availqty BIGINT not null,
-    ps_supplycost DOUBLE not null,
+    ps_supplycost DECIMAL(15, 2) not null,
     ps_comment STRING not null
 ) storage_format = 'native' compression = 'lz4';
 CREATE TABLE IF NOT EXISTS part (
@@ -58,7 +58,7 @@ CREATE TABLE IF NOT EXISTS part (
     p_type STRING not null,
     p_size INTEGER not null,
     p_container STRING not null,
-    p_retailprice DOUBLE not null,
+    p_retailprice DECIMAL(15, 2) not null,
     p_comment STRING not null
 ) storage_format = 'native' compression = 'lz4';
 CREATE TABLE IF NOT EXISTS region (
@@ -72,6 +72,6 @@ CREATE TABLE IF NOT EXISTS supplier (
     s_address STRING not null,
     s_nationkey INTEGER not null,
     s_phone STRING not null,
-    s_acctbal DOUBLE not null,
+    s_acctbal DECIMAL(15, 2) not null,
     s_comment STRING not null
 ) storage_format = 'native' compression = 'lz4';
